Change build system to CMake, cleanup warnings

+++ src/MountProg.cpp
@@ -85,7 +85,7 @@ void CMountProg::Export(const char *path, const char *pathAlias)
 	if (path != NULL && pathAlias != NULL) {
 		if (m_PathMap.count(pathAlias) == 0) {
 			m_PathMap[pathAlias] = formattedPath;
-			printf("Path #%i is: %s, path alias is: %s\n", m_PathMap.size(), path, pathAlias);
+			printf("Path #%lld is: %s, path alias is: %s\n", m_PathMap.size(), path, pathAlias);
 		} else {
 			printf("Path %s with path alias %s already known\n", path, pathAlias);
 		}
@@ -219,7 +219,7 @@ void CMountProg::ProcedureEXPORT(void)
 
 	for (auto const &exportedPath : m_PathMap) {
 		const char* path = exportedPath.first.c_str();
-		int length = strlen(path);
+		int length = static_cast<int>(strlen(path));
 		// dirpath
 		m_pOutStream->Write(1);
 		m_pOutStream->Write(length);
